backend-fastapi/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==41.0.7
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
email-validator==2.1.0
bcrypt==4.1.2
